# docs/swagger.yaml
basePath: /api
definitions:
...
  order.CreateRequest:
    properties:
      ad_id:
        type: string
      company_id:
        type: string
      description:
        type: string
      influencer_id:
        type: string
      price:
        type: number
    required:
    - ad_id
    - company_id
    - influencer_id
    - price
    type: object
  order.UpdateStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: API Server for Marketplace for influences and companies
  title: Marketplace
  version: "1.0"
paths:
...
  /order:
    post:
      consumes:
      - application/json
      description: Creates a new order for a company and influencer
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order created successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to create order
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - order
  /order/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to delete order
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete an order by ID
      tags:
      - order
    get:
      consumes:
      - application/json
      description: Retrieve an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve order
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - order
  /order/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/order.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Status updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to update status
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - order
  /order/company/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all orders for a specific company
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve orders
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get orders by company ID
      tags:
      - order
  /order/influencer/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all orders for a specific influencer
      parameters:
      - description: Influencer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve orders
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get orders by influencer ID
      tags:
      - order
...
securityDefinitions:
  BearerAuth:
    description: JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"